Refactor RedisMetadataRepository and add Spring-specific loader
Refactored RedisMetadataRepository into an interface and moved the Spring Data Redis-specific implementation to SpringRedisMetadataRepository. Added SpringRedisMetadataLoader and service registration for RedisMetadataLoader. Updated imports and usages throughout the codebase to reference the new Spring-specific classes. Added related tests and resource files.
