+++ microsphere-redis-spring/src/main/resources/META-INF/services/io.microsphere.redis.metadata.RedisMetadataLoader
@@ -1,0 +1,1 @@
+io.microsphere.redis.spring.metadata.SpringRedisMetadataLoader
