+++ microsphere-redis-spring/src/test/java/io/microsphere/redis/spring/metadata/RedisMetadataRepositoryTest.java
@@ -21,11 +21,11 @@ import org.springframework.data.redis.connection.RedisConnection;
 
 import java.util.function.Function;
 
-import static io.microsphere.redis.spring.metadata.RedisMetadataRepository.getRedisCommandBindingFunction;
+import static io.microsphere.redis.spring.metadata.SpringRedisMetadataRepository.getRedisCommandBindingFunction;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 /**
- * {@link RedisMetadataRepository} Test
+ * {@link SpringRedisMetadataRepository} Test
  *
  * @author <a href="mailto:mercyblitz@gmail.com">Mercy</a>
  * @since 1.0.0
